Readme.md: Fix 'Bi as' -> 'Bias' typo
I also wrapped the paragraph to 70 characters (Emacs' default [1])
while I was touching it, since just joing 'Bi' and 'as' produced a
line that was over 200 characters wide.  Wrapping *all* the paragraphs
in the README seemed too invasive for this typo fix, but feel free to
recommend an alternative width or wrapping strategy.

[1]: https://www.gnu.org/software/emacs/manual/html_node/emacs/Fill-Commands.html